+++ modules/cms/src/admin/messages/es/cmsadmin.php
@@ -173,4 +173,5 @@ return [
     'cmsadmin_created_at' => 'Creado en',
     'cmsadmin_version_remove' => 'Eliminar versión',
     'view_index_sidebar_container_no_pages' => 'Empty container',
+    'view_update_set_as_homepage_btn' => 'Set as homepage',
 ];