+++ modules/cms/src/admin/messages/ua/cmsadmin.php
@@ -194,4 +194,5 @@ return [
     'cmsadmin_created_at' => 'Створена',
     'cmsadmin_version_remove' => 'Видалити версію',
     'view_index_sidebar_container_no_pages' => 'Empty container',
+    'view_update_set_as_homepage_btn' => 'Set as homepage',
 ];